Refactor strategy modules
